API tested and added some BOOK DTO files and all successfully running on prod.

+++ src/main/java/com/org/bookstore_backend/DTO/AuthorDTO.java
@@ -11,6 +11,6 @@ import lombok.NoArgsConstructor;
 public class AuthorDTO {
 
 
-    private int authorid;
+    private Long authorid;
     private String name;
 }
